git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2274 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkStudio/src/rws/components/luaeditor/CodeEditor.cpp
@@ -143,7 +143,7 @@
  }
 
 CodeEditor::LineState CodeEditor::getLineState(int lineNr){
-	if(_executedLines.size()<=lineNr || lineNr<0)
+	if((int)_executedLines.size()<=lineNr || lineNr<0)
 		return Nothing;
 
 	return _executedLines[lineNr];
